feat: Show the hit chance for the Phial of Floods
The Phial didn't show its hit chance in the targeting UI due to the
unique setup it has. Ideally it would use your_spells() for most of its
setup, like wands and the Lightning Rod do. Other than using a custom
range that differs from the default range of the Primal Wave spell, I'm
not sure if there are reasons why it can't use your_spells(), so I'm
just modifying the existing approach for now.