feat: Evokable accuracy info for clua
Add a target_evoke() method to the monster info table. Like the
corresponding target_weapon(), target_spell(), et al. methods,
target_evoke() gives the accuracy string seen in the UI when targeting
the given monster with the given evokable. It expects an item obtained
from the clua items library as an argument, and only returns a non-empty
value for wands or the Phial of Floods. The other evokables either
always hit or otherwise don't have a meaningful notion of accuracy.