feat: Add damage and noise to evoker descriptions
For the XP evokers that do direct damage and have casting or effect
noises, namely the Tin of Tremorstones, Phial of Floods, and Lightning
Rod, show the current damage die and noise values in the item
descriptions, like we already do for wands.

The Lightning Rod damage situation is a bit complicated due to arc
length decreasing the damage, but we mention this now in the main
description text.