feat: a webtiles lobby idle timer
Over time a running webtiles server tends to accumulate lobby
connections where a browser window is still open to the lobby after
someone has saved; CAO runs at about 40-50 at any given time (a big
proportion of the total sockets), and I tend to think the vast majority
of these aren't real. This commit adds a basic idle timer running in the
lobby that is configurable (and disableable) by the server admin. At the
end of the timer, the connection is closed. Any action in the lobby
restarts it, but there aren't many actions that can be taken. The
default is enabled with a fairly long (3hr) timeout.

TBD: does this break beem without further special casing?