fix: levelgen freeze with zin + randart talisman
`@god_name@` in a randart talisman name would trigger a check as to
whether the randomly chosen god likes the artefact. There was an
accidental dependency in this code path on you.religion specifically for
god_despises_artefact, so the wrong god would be checked for this case,
and in particular a zin character would cause levelgen to run in a tight
loop while trying to generate a talisman. (Possibly the relevant code in
`replace_name_parts` should have an escape clause of some kind?)