Enable a confirmation prompt for "bump" attacks with ranged weapons.
If the user attacks by moving towards a monster holding a ranged
weapon, this change causes the player to be asked to confirm the attack
in some situations.

A "!a" inscription might prompt a "Really attack ...?" message, which
is temporarily suppressed after a "yes" response. A "!f" inscription
might prompt a "Really fire ...?" message which appears for each
attack.