Rework monster speed descriptions
Speed is one of the most important traits a monster had, but we buried
it in a sea of trivia and described it quite vaguely. The old speed
adjectives used to work OK, but once we started adding monsters with
extremely fast or slow actions (e.g. juggernauts, animated armour),
they were quite inadequate.

Instead, show:
- A speed percentage (eg 150% for 15 speed monsters, 80% for 8-speed mons).
  This percentage is unaffected by haste, slow, etc - it's just base speed.
  Higher numbers are better. Hopefully this is intuitive.
- More percentages for special action types ("swim: 166%", "attack: 33%").
  These percentages are pre-multiplied by base speed (so a juggernaut with
  33% attack speed takes about 3.0 dAut per attack), and higher is still
  better. Hopefully this is intuitive.
- An estimate of whether the monster is normally faster or slower than
  the player, without taking into account buffs/debuffs or terrain for
  either the player or the monster. There might be some int rounding errors
  here.