Remove size's effect on shields (again)
Continuing from 6e8f0f562f661212d597, remove the effect of monster
size on SH. (For simplicity and sanity.)

This is a small nerf to spriggan monsters, which get -4 SH now (if
they had bucklers before). Spriggan riders, for example, go from 24
SH to 20. We can comp 'em some HD or EV or something if needed.