Tweak glass eyes' description
57150269e139b239e5 made glass eyes LRD-able, based on an early WIP
description of glass eyes that described them as being literally
glass. In the meanwhile, the final version of 492612b1e5dc9033738
changed glass eyes to only be glass-y. This was inconsistent.

The mechanical implications of the decision are limited, and player
intuition probably does want 'glass eyes' to be literally glass.
Adjust the description to match, and cut out a sentence that
duplicates the 'spells' section.