New Mines enemy: kobold bombardiers
Part of how Orc got its levels cut in half was the lack of variety inherent
to an orc-centric branch, with a variety of orcs already sitting on most
of the possible roles orcs could play, and with the obvious other cavedwellers
mostly overlapping with the rest of the game. This is part of an experimental
batch of new enemies meant to fleshen out the identity of the Orcish Mines on
the latter part rather than the former, which should also grant it a fair bit
more enemy variety to actually vary up its ends with beyond the standard four
strongest orc types and rare per-end uses of ogre / centaur / giant guests,
all already available elsewhere in D and V.

Kobold bombardiers are uniquely both ranged monsters and conjurers,
benefitting from the distance of the former and the burst damage of the
latter. They use the player species' +3 ranged weapons apt with an orcbow,
arbalest, or very rarely a hand cannon, and exaggerate this further by also
operating a monster-only cannon with which they slowcast Bombard. The latter
deals quite fairly more damage and lets them back away from those approaching,
but is more limited in range compared to their weaker arbalests, which should
provide an interesting tension without needing too much further complication.
(Possibly monster Bombard might warrant an 100% self-knockback rate rather
than a 50% self-knockback rate, to be any noticeable compared to Iron Shot,
but I don't know if that'd be a bit too much of a buff to ghosts using it).

They naturally spawn in duos, and probably could fit in a slim number of D
vaults already focused on kobolds. A couple of less conventional Orc end
vaults (like the one currently reliant on centaurs, the serial vault towns
and mushroom circles, and the encompass vault) will get these in the future.
For now, they replace the completely harmless gnolls in the branch alongside
a little bit of the orc warrior bands and cyclops weight, skimming off of
monsters one sees plenty of in D. They're more dangerous than any of those,
but they're probably not more dangerous than higher-tier orcs, so it should
be fine.

(Extremely placeholder tiles frankensteined between roctavian's old big kobold
tile, needle trap tiles, and one of the granite statue tiles. TODO: decide on
the best name between kobold demolitionist (maybe too close to demonologist?),
kobold blastminer (extremely on the nose but maybe the best reading?), and the
bombardier name used here (less rather direct and not a portmanteau?))