adjust a zig pillar and zig set order (#3359, #3364)
[Commiter's notes: Lowered the pan lord chance even further, so that
there's less bias against Shatter for dipping the early floors of a zig
in normal 15 rune games. The order of hells listing is somewhat arbitrary
but can be made to match the pure element sets right after it, so I'm
reorganizing the monster list order to match rather than the comments.

Closes #3359 and #3364.]