fix: better handle webtiles ttyrec write errors
If a ttyrec write on game start triggered an OSError (e.g. on a full
disk), the process object wouldn't get started, but also wouldn't get
fully cleaned up due to a bug, resulting in spurious processes in the
lobby. This fixes this bug, so now the process should get cleaned up
(and the user gets an error message).

This commit also tweaks the handling of OSError on ttyrec writes for a
started game. Previously they just led to an unhandled exception from
the callback, now they log a better error; unclear if something else
should happen.