Make living jellies immune to nets, note this + their rConstrict in x-v
Making jellies immune to nets is consistent with their immunity to
constriction, and matches with how insubstantials are immune to both.
It's a nerf to throwing nets, but we just buffed nets mildly by being
able to hit SIZE_GIANT monsters in 7dc041b anyway.

Aside from making fused slime creatures a bit scarier to run from, this
is most noticeable for removing one of the many TRJ strategies. There's
possibly a little concerning increase in threat in Slime between that, some
of the newer slime end layouts, and the new prevalence of glass eyes and
their non-golden / creature company. Might be worth looking into with
slightly reduced floor spawns or something, but it's probably fine for now.

This also adds a note to x-v that jellies are amorphous and immune to
ensnarement (both nets and constriction), as is noted for insubstantial
monsters. It'd be nice if this amorphousness note also was useful for
Shatter's effectiveness, but Shatter checks the general Jiyva creatures
and I don't think eyes should get immunity to either of the ensnaring
effects? Probably.

(XXX: jelly simulacra being too solid to avoid nets / constriction is fine
from a flavour perspective, but does that justify jelly zombies also being
vulnerable to the same stuff?)