Make monster Bombard knockback the caster 100% of the time
Currently the only users of it are player ghosts, and player ghosts being
optional weird threats with dramatic spells mostly assigned to them is
their current entire point, so them getting Bombard as opposed to Iron Shot
should be more noticeable. It also makes any future monsters (as is slated
in a future commit) casting Bombard be more parseable and consistent at
using that as a gimmick mechanic compared to the normal Iron Shot.

(It's not symmetric with player Bombard still being a coinflip to move,
but monsters have lots of ways their spells are stronger than players
anyway, and nobody's complaining about monster summons working out of
monster LoS currently.)