Add controversial Rakshasa tile (Sastreii)
The depiction of rakshasas as tigers seems to be an antique D&Dism.
To be mythologically accurate, we should move away from it; or, if
we want to keep this sweet art, we should retheme rakshasas as eg
'felid illusionists' or 'devil-cats' or similar.

Anyway, this is newer and more dynamic than the previous cat-like
Rakshasa tile, so let's use it :)