New species: Coglins
Wielding weapons in each hand, Coglins are puny goblins that ride magic
exoskeletons into battle. They're slow and inflexible in many ways.
When they stop moving, their exoskeleton will be slow to move again.
When they stop swinging, their exoskeleton's strikes will be weaker until
they rev back up again. And when they switch weapons, it takes them some
time to re-attune their exoskeleton's spirit. But once they get going -
look out!

More notes on mechanics:
- Dual wielding: you can choose to wield a 1h weapon instead of a shield
  in your off-hand. (Unless you're wielding a 2h in your primary hand.)
  Whenever you strike with your primary weapon, your secondary will also
  attack. This doesn't trigger extra aux attacks. You can dual-wield 1h
  ranged weapons, but mixing and matching ranged and melee weapons doesn't
  work. Delay is the average of the two. The order of attacks is random.
- Warmup strides: whenever you take a turn that's not movement, your next
  movement will take twice as long as normal. Other actions (attacks,
  spells, etc) remain normal speed.
- Warmup strikes: your ranged and melee attacks are 33% weaker until you
  fully rev up, which takes about 4.0 dAut of attacking enemies.
- Slow wielding: wielding, unwielding, or switching weapons (in both the
  primary and off-hand) takes 5 standard turns instead of 0.5.

The idea is to play in a similar space to formicid's 2h+shield ability.
Dual wielding is expected to *usually* be stronger than wielding a 2h or
1h+shield - and if it isn't, we can buff it until it is - but gear
found will sometimes prompt one to use a shield or a 2h weapon, for a
shorter or longer period. That sort of 'soft' change for gear evaluation
(as opposed to a 'hard' lockout on item types ala spriggan/kobold/ogre/etc)
seems fun and exciting.

The various downsides are intended to complement this. Dual-wielding is a
big strategic upside, so the downsides are tactical, hopefully making fights
feel a bit different without being overly obtrusive. (And slow-wielding, in
particular, is intended to avoid excessive weapon micromanagement.)

There's also an implementation of an unused 'attractivitis' mutation in this
commit. We may later turn this into a generic badmut, or cut it entirely.
downside, either!

More notes:
- Weapon names are purely cosmetic, but Esmeralda's feelings will be hurt if
  you say so.
- Apts and attributes are early drafts and especially subject to change.