Enhance max enchant messages
Change what happens when weaponry or armour in enchanted to its
maximum.

1. Print "while" rather than "moment" in the message, to give the
player a hint that something unusual has happened.

2. If caused by a scroll, cancel repeated actions. The next scroll
can't enchant this item again, but the same keypresses could enchant
something else.

Also, print the message at the end of enchant_armour(). This matches
enchant_weapon(), and means that any lua function which is triggered by
the message sees the item in its final state.