Remove Me(teoran)
Following on from the last two commits, Gems now serve the purpose
that Meteorans were intended to, and Humans have taken their healing.
Meteorans can now fly on to another star. Farewell, sweet princexes!

I have a new species to fill Meteoran's "slot" in-progress, but the
current planned 0.31 release is too close for it to land yet. More
on this in the next few months, with any luck.