Remove explore healing from Arena
Folks have noted for a while that Meteoran gets a huge burst of
explore healing from entering Okawaru's Arena. This is kind of cute,
but it's not really exploring, and it's not in keeping with the
intention of explore healing. Since Arena is quite strong enough as
is, let's cut this. (Don't think this will violate player intuition.)