Brain Bite adjustments and fix-ups
Make the intended zero-mp check into a low mp check so monster randomness
doesn't struggle against natural player MP regen (with the threshold being
the 20% max mp it takes away). Display a different message and move Xom
interest to when extra damage is being dealt. Adjust the damage from 4-12
at the sufficient mp threshold to 8-16 so the description can just say it's
doubled from a 4-8 base. Display that base with an asterisk in the in-game
and out of game monster look-up, as LRD currently does, since most of the
time one's MP should be above the doubled damage point.