Make red very ugly things not lose AF_FIRE for AF_STICKY_FLAME
AF_STICKY_FLAME made sense as an upgrade mostly when there was still item
destruction; when we lost item destruction in 0.15, sticky flame's
non-scaling damage made it pretty harmless beyond pretty early parts of the
game. Sticky flame could do with a new mechanical purpose, but until it
gets it, it is probably for the better that something so prominent in
3-rune winning games as 19 seperate spawns isn't noticeable weaker than its
default form if it rolls the wrong element.