Cloud card: bring back select other cloud types
3194668 removed the variable cloud types from the Cloud card because the
damage options made it inappropriate for the deck it was in. While this is
a reasonable direction for it to follow, retaining the cloud-ring aspect
for harmless if opaque clouds mostly just makes it a very awkward version
of scroll of fog, for being centered on enemies instead of the player.

This re-adds different cloud types to the cloud card, but each tier uses
different types of debuff clouds rather than the old set of clouds including
poison, fire, ice, or acid clouds.

 * L0: mephitic fumes
 * L1: miasma
 * L2: calcifying dust.

Also, in the same spirit of encaging enemies it previously had, it once
more doesn't place directly on monster tiles (but now actually works for
non-wall tiles like fountains and open doors like it should), and they
also last less time than they used to. The later two clouds don't overlap
with condenser vane, and while there's some spell overlap the cloud ring
functionality making many monsters not move forward but getting debuffed
when they do should be a reasonably distinct effect compared to the
limitations of Cigotuvi's Dreadful Rot, the will-check of Petrify, and the
cloud lines of Elements card catoblepas summons.

(Possibly Nemelex should reward very slightly less cards if any more cards
get revised at this point, but escape cards are already usually just
stacked or triple-drawn from outside of pretty earlygame, and the sheer
power of Tomb, Elixir, or higher-power Exile still look much better off
than this.)