Tomb card: use non-default wall tiles
When using the Tomb card, in sufficiently messy terrain or with lower power
leading to incomplete wall raising, it can be easy to forget which walls
will be the ones that'll eventually go down unless one bothers to check
over each surrounding wall. This ui quandry is due to it using the given
branch's default rock tiles, so over in Tiles it can be made clearer just
by using a distinct rock tile- I've chosen WALL_SANDSTONE for this, since
it's what's used in wizlab_dorokhole.

(Note to self: wall_sandstone is inconsistently used for rock or stone
in different vaults, should fix that to the former.)