Don't blame spriggan druids for their friends
We usually describe the ikiller as a band leader or summoner -
a creature with pretty direct responsibilty for its pal who just
killed you. Including spriggan druid convoking is a stretch. Should
ironbound convokers get blame, or guardian serpents, or any number
of buffing monsters..?

Basically, it makes the stats very hard to parse, which wasn't the
intent.