Print monster names for unseen shouts
So players don't have to look up what on earth makes "a rustling",
or what the difference between a skittering and a faint skittering is.

Might give too much info for shout luring. I've disabled it for some
common shout types, but not sure if that's right, or if I should
disable it for more...