fix: checks for chaos flaming and freezing (hellmonk)
A similar issue to that fixed in 831ba139. Use the correct checks for
allowing these brands from chaos. The current checks looked for "fiery"
and "icy" monsters respectively, which does cover a bunch of monsters
immune to fire damage, but not all of them. Instead we check that the
target is either the player (who can never be fully immune to
flaming/freezing) or, for monsters, if it has less than 3 levels of the
relevant resistance, otherwise it would be immune.