Add gems
When I added Meteoran (1352289c90d, based on Hellmonk's ad05b8d819def),
I wanted to give players a fun way to engage with time pressure. When I
play, I really enjoy the feeling of exploring on less than full HP and
making choices about which areas to explore. (Full clearing everything
carefully can be fun, too, but variety is the spice of life!) I'd hoped
to find a way to bring that playstyle to the wider playerbase, with time
limits that are more defined and achievable than the harder and fuzzier
goal of 'compete for a high score'.

Some people, including myself, really liked Meteorans! But a large
number of players, probably the majority, found them stressful and unfun.
That isn't the end of the world - I'd much rather have a species which
20% of players love and 80% hate than one which 99% of players don't care
about one way or another. But I'd also hope that we could do better.

Gems are intended to be an alternate approach to the 'time pressure'
playstyle. They're a new type of collectible item, appearing at the end
of DOLESAPNMVCWUZ - basically, all the non-portal, non-temple branches of
a 3-rune game. Each gem has an associated time limit, which ticks down
while the player is in their associated branch. When that time runs out,
regardless of whether the player has gotten the gem yet or not, Zot rudely
smashes the gem into ten thousand pieces.

The intention is to allow several different ways to interact with gems:
- Ignore them, or not even realize they exist. This is intended to be the
  primary/default interaction.
- Dive to grab gems, but not bother trying to keep Zot from smashing them.
  This is a 'lightweight' speedrunning playstyle, a bit like the Speed
  Demon I tournament banner.
- Keep one or more gems intact by only exploring part of a branch, perhaps
  opting to 'go for it' when your character is feeling especially strong.
- Try to grab all gems and retrieve them all intact. This is roughly
  equivalent to the old Meteoran playstyle.

Gems have absolutely no mechanical use within the game. They offer a very
minor score bonus (10k points each), as a bonus for new players who look
at scores for unwon games, but shouldn't affect normal play or speedrunning
in any way.

Getting the Orb of Zot shuts the timer off. One could skip branches, get
the orb, and then clear branches while holding the Orb to get gems with no
time limit... but orbrunning is its own form of time pressure, and I'm
skeptical this would be easier than playing 'normally'. :)

Time limits are currently set on a per-branch basis. Lair, Vaults, and
Depths have longer time limits than they did for Meteoran, to allow for
large levels and travel time, while Slime and Zot are shorter. However, I
would be startled if these time limits stuck - I personally suspect most
of them are too tight for species which *aren't* as strong as Meteoran.
We can experiment.

The good gem tiles are by Sastreii, the bad ones are by me. :)
Credit also to ellpitic for helping to set me down this path.