Make dream sheep headbutt, not bite (#2740)
The Github issue in question seems to want to add a new AT_HEADBUTT, but we
seem to have made AT_HEADBUTT for... ugly things having random verbs... for
over a decade without using it directly anywhere else. It does seem to be
very easy to find videos of sheep headbutting people and one another and
other animals, while biting seems to only happen when they're overly feisty
while being handfed. So.

Closes #2740.