Shove in two more player -> monster spell equivalents for illusions/ghosts
Necrotise translates to Pain, Confusing Touch to Confuse. Both are
imperfect, but we've got a few examples of that already with Dragon's Call
-> Summon Dragon and Swiftness -> Sprint, and they're reasonably close.
Also, the current player to monster gap for starter background spells for
Illusion Card and player ghosts is incredibly high, so a little less
pressure would help a bit- this should get a little more attention in some
eventuality, so both don't drift off nearly as much.