Let monsters get Dispel Undead (non-Range)
When Dispel Undead Range was created to split the melee and range versions
between players and monsters, the former wasn't put into mons_spell_beam.
This prevented monsters from getting it, which wouldn't normally be a
problem since monsters are mostly given Dispel Undead Range by default.
Player ghosts and player illusions, however, are supposed to get the
player's version of it, and since that works, we might as well give them
the shorter range version of it.

The hit dice of this can get fairly obscene for ghosts, but since it's not
a starter spell for necromancers and player ghosts were shoved into runed
door vaults, it shouldn't be too concerning for Mu / Vp / Gh players to
suddenly explode without at least some reasonable counter-measure
assessment. Should be a good sight in negative energy zigs and continue to
help fix up Illusion Card a little more, at least.