Two quote additions, some quotes.txt clean-up
The new quotes are for abyssal stairs and the Petrify spell. The
formatting for either source might be a bit off, since we're inconsistent
about bothering with sections, years, or numbers for webcomics- if anybody
feels strongly about this, they should also go update the Order of the
Stick quotes alongside these two.

 * Potions of blood, sling bullets, and bolts all don't even have item
   descriptions, so they don't need quotes. Arrow quote now is used for
   Poison Arrow. (If we get new tiles and underlying code bothering to
   make hand cannons fire bullets, probably they could use the sling
   bullet code, but currently they fire bolts.)

 * Wizard quote has been given to arcanists and occultists.

 * Storm Form and Necromutation quotes have been put under the matching
   new talismans.