Revamp Call Canine Familiar
The fact that CCF randomly rolled between different tiers of preexisting
canine based on player spellpower had always created some awkwardnesses.
The power jump between tiers could be dramatic, and when the summon
duration was higher, could even encourage rerolling to get a better dog.

This replaces the reliance on tiers of existing monsters by introducing a
new (spell-exclusive) canine that can instead scale smoothly in damage,
health, and AC based on player spellpower, without any breakpoints. No
more rerolling: now you always get the dog that you deserve.

But there's more! In an effort to have more 'interactive' summon spells
that don't simply fall into the hole of 'also throw a dog at it' along with
everything else you're summoning *and* emphasize the familiar bond more,
the spell has the following additional behavior:

-If your inugami dies, you are drained and cannot resummon it for a short
 while.
-Recasting the spell while your inugami is still active, instead *heals* it
 and causes its next attack (within a few turns) to cleave and cost no
 energy.

It should be individually a small bit stronger than the old dogs, and
encourage more interacting with it, to make best use of the healing and
cleaving, based on the situation.

All numbers associated with this ought to be vaguely reasonable, but are
still subject to change.