Don't cancel Awaken Forest the moment you swing in a druid's direction
More specifically, the game was calling mons_att_changed() whenever the
player performed an offensive action at an enemy - even if that enemy was
already hostile. This was often harmless (if unexpected), but it did result
in two strange behaviors:

-Attacking a druid (even if you missed!) would immediately end its Awaken
 Forest effect as though it had changed sides.
-Attacking a tentacle (or tentacle head, like a tentacled starspawn) would
 cause all its tentacles to stop constricting anything they were
 constricting.

This attempts to only call mons_att_changed() after attacks if that monster
was not already hostile. (But behavior code is complex and I can imagine
there being some other scenario this misses.)