Awaken Forest tweaks
Adjust the duration to be a little more consistent (higher minimum duration
and a greater tendancy to be closer to the average).

Also change the cast logic. Previously they would only cast Awaken Forest
if at least one enemy was *already* adjacent to a tree. In practice, using
it if the player is *near* a tree in Swamp is still likely to constraint
future movement options in an interesting way, so now it checks if any
enemy is 2 tiles away from a tree instead.

I also raised the cast rate of it for druids somewhat - at least partially
to make up for 'gambling' that a cast will hit the player in future and
sometimes behind wrong. (Since they won't recast while it's active, this
should waste minimal turns).