Fix broken inugami see invis
They already had the M_SEE_INVIS flag and it was never intended that they
lose it compared to existing dogs, but apparently if monsters have a
ghost_demon struct, the game looks there and only there for whether they
should be able to see invisible.

ghost_demons are weird.