fix: don't crash on call canine familiar
Repro instructions (Implojin): drink !invis, cast call canine familiar.

Prior to this commit, the game would crash if an inugami `ghost` was
accessed by side effects of `create_monster`, which was possible (e.g.
if there was an invisible actor nearby, a behaviour_event was triggered
immediately before the return to the casting code). Generally,
ghost_demon code assumes that the ghost gets fully set up in
`create_monster`. This commit initializes a ghost object there at least
so that it won't crash (along with a bit of tweaking of the starting hp
values), and then finishes initialization in the spell, where there is
access to spellpower. I'm a bit concerned about spreading this out being
somewhat brittle (I don't think any existing ghost demons do this so
there may be violated assumptions), but it seems to work in some quick
testing.