Use a PromptMenu for various selections in wizard mode.
Various wizmode functions functions printed a menu in the message
window, but didn't check if it would fit, and thereby hid some of the
options.

Make some use PromptMenu instead. This displays the menu in a separate
popup if it's larger than the message window, so doesn't share this
issue.