Leda's Liquefaction experiment: Earth / Hexes to Earth / Tmut, + buff
Leda's Liquefaction started as Earth / Transmutations in 0412296. It was
changed to Hexes as part of the double-combo of ae9be96 and 06cfd5e moving
it to Hexslinger so the Hexes-based ranged background could focus on and
cast something they're particularly suited to use. 9e18931 moved it out of
the starter book because a level 4 dual-school spell split a weapons
background too many ways, and ever since then it's been left pretty wanting
and unused. It's inherently a rather risky spell, as is anything that slows
down the player's movement, and doesn't particularly fit most of Hexes'
desire to run up and stab things nor is it necessarily likely to tempt
ranged characters at its weird schools intersection point. It's gotten
some buffs, but is still an unpopular fit.

Changing it from level 4 Earth / Hexes to level 3 Earth / Transmutations
covers a fair number of different bases. Hexes is extremely packed in its
lower-level and mid-level spells with the recentish Hexslinger update, so
it could afford to move a spell to another school. Transmutations as a
school losing forms in 1880023 hasn't completely killed the school, since
the unarmed focus rarely intersected too much with the power of Ignite
Poison / Irradiate / Yara's, but it is definitely a school that could do
with more unique effects on its own even if as a flavour grab-bag of
non-Hexes buffs / debuffs / damage. Moving it to Transmutations also pushes
it away from the school containing direct Slow spells, while it works on
a very different vector anyway. Finally, I think Transmutations always
made a lot more sense- there's no Willpower check or HD check involved, and
it clearly is transforming the earth near the caster anyway.

Leda's Liquefaction also gets a mild increase of base duration outside of
power (10 -> 15) to help it out a little more, since Transmutations itself
is rarely a school and skill to contribute spellpower with, and since it's
already a risky spell without needing to be ineffectual when actually
tried. In terms Placement-wise, not much has happened beyond moving it from
the Book of the Hunter to the Book of Sloth. I'm in two minds about
possibly moving it to spell level 3 if this still isn't enough...