A quick tuning down of Xom's banishment power (Lici)
Xom's interest levels restrict the chance to banish the player, but
currently, Xom's banishment power uses player xl as a stand in for monster
banisher HD. While there's checks before xl 9 to have Xom usually rescue
the player, this scaling works pretty poorly between what power we expect
of players at given xls versus given monster HD- for example, deep elf
sorcerers have only 14 HD to send the player down to Abyss:3, which is
reasonable at the time one's doing Elf but not so much for an end-of-Lair
character that happens to be xl 14. The current range looks like:

xl  | 1   | 2-6 | 7   | 8-12 | 13  | 14-18 | 19  | 20-24 | 25  | 26-27
-----------------------------------------------------------------------
min | A:1 | A:1 | A:2 | A:2  | A:3 | A:3   | A:4 | A:4   | A:5 | A:5
max | A:1 | A:2 | A:2 | A:3  | A:3 | A:4   | A:4 | A:5   | A:5 | A:6

While there's a fair bit more abyssal agency than many years prior, this is
still quite agressive even for Xom, considering how much later it can ruin
a character. I'm doing a quick formula adjustment to the player's xl in the
equation to make this more reasonable for now. The resultant possible
minimum and maximum depths now look like:

xl  | 1-11 | 12-15 | 16  | 17-20 | 21  | 22-25 | 26  | 27
-----------------------------------------------------------
min | A:1  | A:1   | A:2 | A:2   | A:3 | A:3   | A:4 | A:4
max | A:1  | A:2   | A:2 | A:3   | A:3 | A:4   | A:4 | A:5

Possibly we could make people in zigs or with 5+ runes visit Abyss:6,
but people playing extended Xom probably would manually go to Abyss:7
themselves anyway. Should also look into Lugonu being less boring...