Increase the number of inventory slots to 104.
Expand the inventory to 104 slots. The first 52 things you pick up are
assigned to letters; the last 52 are all assigned to &. Items can be
used (or reassigned) from the inventory list.

The game doesn't, by default, try to use the same slot if you drop a &
item and pick it up again.

Also enable the ; (last piece of equipment to be removed) shortcut to
work in any inventory menu.