Buff Jiyva MP regen to match the amulet (acrobat)
This is the same size buff that amulets of MP regen got in 12fdbb3a, so
now the HP/MP regen ratio under Jiyva is again the same as between the
respective amulets (so at max piety Jiyva regen boosts are the same as a
free amulet of Vitality). I'm somewhat wary of making too much MP regen
available in general (there's a reason why MPRegen isn't a randart property
and HPRegen is), but on a non-temple god like Jiyva it seems fine to go
a little crazy.