Don't drain the caster when their inugami dies
This was originally intended to incentivize keeping your familiar alive,
but the recast cooldown probably does a sufficient job of that in the
earlygame on its own.

More importantly, this interacted extremely poorly with guardian golem,
repeatedly punishing the player for ever daring to use them together. While
I fully intend to do something with golem itself in the near future, this
should help smooth over some bad gamefeel in the meantime.