Juggle around the unused monster tiles folder some more
Considering a fair number of the tiles will forever exist in the git
history, rltiles/UNUSED is a weird batch in an untidy closet we're
evidently never going to bother to fully clean out. There's tiles in here
that are in active use at the current moment, older versions of currently
present monsters (fine for former-humans, baffling for generic demons), and
a variety of tiles that wouldn't really past muster for any even short-term
use. Whatever gets here versus entirely removed basically boils down to
random chance based on a dev bothering to keep given tiles or not.

This brief tour cuts out dozens of redundant or mediocre tiles left lying
around, alongside tiles currently in use one way or another. It also adds
back a handful of tiles that look any decent, to help out with making
future placeholders and to help out other works using the rltiles CC0
license. (It adds a bit to the repository, but crawl's tiles are so so tiny
that it adds no real burden compared to e.g. adding whatever new splash
screens we get.) Re-added tiles here include coolio's spirit and griffon,
roctavian's death ooze / ravenous mimic / giant mite / ants, and Denzi's
rock worm. Also, all the non-tall lord / Tiamat tiles are now consistently
labeled as .small.

If anybody really misses some of these tiles, maybe we could do with a
subfolder for this that contains every tile we've replaced or removed,
seperate of the main folder.