Small Xom overflow and Orc decor vault batch
The Xom vaults were inspired by 6d556a0 and 292434c. Probably the 0.2%
chance for Xom to present completely unusable randarts is fine, alongside
the 2% chance for a pile of box of beasts mimics.

The Orc decor vaults are all somewhat inspired by Orc ends in various
fashions and are meant to help a little bit with making an Orc decor
guarantee, alongside the additions from PR #3325 whenever that's done.