Buff and shorten Vitrify, update UI around it, spread it a little more
Vitrify seems to have been relatively well recieved for the prospect of
having a new dangerous Hex beyond paralysis and banishment, but there's
been relatively low fears of it due to the conservative rate it started out
at. I'm buffing it from +30% damage to +50% to make it match the damage
numbers of negative resistances and be dangerous enough it has to be
worried about immediately. The gaze goes from increments of 6 to 10 turns
to 4-8 turns, and the spell goes from 8-24 to a less harsh cap of 8 to 18.

To correlate with this buff, the default messages settings now make players
being vitrified get the same danger colours as constriction and corrosion
in the message log. I'm also adding the exact damage increase it inflicts
to the % screen, to correlate with the exact numbers currently displayed
for harm. There was a fair bit of divided opinion in multiple places
regarding percents like Harm currently uses versus multipliers instead,
or if it should be be made additive instead of multiplicative with Harm.
There's little agreement on the latter, but the biggest opinions were to
display it like current Harm but seperately of the full indicator, a
"+50% damage incoming" which kinda lies if one also has Harm active. I
sort of prefer the multiplicative approach, parsing and expectations-wise,
but if other devs feel strongly about rAll- being additive instead of
multiplicative (or alternatively, if the display should multipliers
or combine the damage incoming together or whatever), they may do as they
wish; I cede the ground to them.

Since the previous state of the effect worked out fine prior, it should
also be reasonable to share it with more than just a few paralyzers:

 * Cacodemons lose the boring Slow and Confuse combo to cast Vitrify and
   a bit more Bolt of Energy. It sort of fits the matter / flesh shaping
   theme Dig and Malmutate already had, and lets them cast a flashier hex
   for greater demon summoning / Pan than spells present the whole game
   long. Maybe their Bolt of Energy could be looked at next for being so
   plain...

 * Klown pies swap out Polymorph in their effects list for Vitrify. There's
   been some consternation about melding the entirety of one's armour being
   far more lethal than most other possible effects around between AC and
   resists, and this makes one still reasonably frail while both keeping
   one's built-up-all-game-long defenses / resists and varying once more
   from the hex or the gaze by checking EV and LoF instead of Will or LoS.