Decouple the Plutonium Sword from transmutation miscasts
So that the miscasts can be removed entirely.

This also adds back a very rough approximation of the direct damage that
the miscasts used to inflict before the Big Miscast Simplification made it
purely cause malmutate/contam.

I kind of miss the tiny chance of mutating random cockroaches from the
floor in old versions and am kind of tempted to give a slim chance to
make a neutral radroach scurry away from you sometimes. But that is for
another day (if ever).

(I also fixed message logic somewhat. Did you know that you'd get a message
whenever the sword triggered, even when it was wielded by a monster
fighting out of los of you?)