Remove orc gem
As with other gems, the orc gem was intended to be an interesting
challenge in itself and to also contribute to the difficulty of the
'all gems intact' challenge. However, Orc's layout meant that players
would often reach the gem quite quickly, even in normal play - not an
interesting challenge. Tightening the time limit would mean that Orc
was a bit more inaccessible than intended. As it was, travel time
to/from Elf and shopping could already create heavy pressure on the
timer.

Let's cut the glittering gem for now. This makes the 'all gems'
challenge a bit easier, and encourages resting in Orc instead of in
Elf:1, but that's liveable. (And similar to other cases re: resting
in Temple, etc.)

It could be interesting to instead make the Elf gem include the Orc
branch in its timer, so it has a longer overall timer but ticks down
while the player is in Orc as well as Elf. That has too high a UI
and complexity burden to be justifiable, though, probably.

Trivai: one now reaches 27 by adding 15 runes + 11 gems + 1 orb.